[Add]_02/Invasion_Controller
몬스터 침입 구현

+++ Assets/Scripts/Eun/Monster_Status.cs
@@ -19,7 +19,7 @@ public class Monster_Status : ScriptableObject
     [Header("Status")]
     public Monster_Type monster_Type;
     public int hp;
-    public float attack;
+    public int attack;
     public float speed;
 }
